% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_built_environment.R
\name{get_built_environment}
\alias{get_built_environment}
\title{Rasterizes buildings and roads onto a common template to create a surface
representing areas that are impassable, preferential walking routes, and
unrestricted areas.}
\usage{
get_built_environment(
  buildings,
  roads,
  template,
  road_travel = 1,
  non_road_travel = 1
)
}
\arguments{
\item{buildings}{A \code{\link[terra:SpatVector]{terra::SpatVector}} containing building polygons.}

\item{roads}{A \code{\link[terra:SpatVector]{terra::SpatVector}} containing roads and paths.}

\item{template}{A \code{\link[terra:SpatRaster]{terra::SpatRaster}} defining the extent, resolution,
and coordinate reference system of the output raster.}

\item{road_travel}{Numeric friction value applied travel along roads and paths.
This is a scalar adjustment to slope-derived Tobler walking speeds.
A value of 2, for example, means that the speed of travel is halved.}

\item{non_road_travel}{Numeric friction value applied to areas that are not
buildings or roads but remain passable.
This is a scalar adjustment to slope-derived Tobler walking speeds.
A value of 2, for example, means that the speed of travel is halved.}
}
\value{
A \code{\link[terra:SpatRaster]{terra::SpatRaster}} representing the built-environment friction
surface. Building cells are \code{NA} (impassable), non-road areas have the
value specified by \code{non_roadpace}, and unrestricted cells have a value of
\code{1}.
}
\description{
Rasterizes buildings and roads onto a common template to create a surface
representing areas that are impassable, preferential walking routes, and
unrestricted areas.
}
