% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{get_dem}
\alias{get_dem}
\title{Downloads elevation data for an area using \code{elevatr} and prepares it to
match a specified raster template.}
\usage{
get_dem(area, template)
}
\arguments{
\item{area}{A \code{\link[terra:SpatVector]{terra::SpatVector}} defining the area for which elevation
data are required.}

\item{template}{A \code{\link[terra:SpatRaster]{terra::SpatRaster}} defining the extent, resolution,
and coordinate reference system of the output raster.}
}
\value{
A \code{\link[terra:SpatRaster]{terra::SpatRaster}} containing elevation data aligned with
\code{template}.
}
\description{
Downloads elevation data for an area using \code{elevatr} and prepares it to
match a specified raster template.
}
