% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_onesource.R
\name{get_target_onesource}
\alias{get_target_onesource}
\title{Transforms a target shapefile into raster format.}
\usage{
get_target_onesource(target, area, template)
}
\arguments{
\item{target}{A \code{\link[terra:SpatVector]{terra::SpatVector}} defining where people are travelling to.}

\item{area}{A \code{\link[terra:SpatVector]{terra::SpatVector}} defining the area of analysis.}

\item{template}{A \code{\link[terra:SpatRaster]{terra::SpatRaster}} providing the extent, resolution,
and coordinate reference system for the output raster.}
}
\value{
A \code{\link[terra:SpatRaster]{terra::SpatRaster}} in which target cells have a value of \code{-1}
and all other cells have a value of \code{1}.
}
\description{
Transforms a target shapefile into raster format.
}
