% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_travel_time.R
\name{get_travel_time}
\alias{get_travel_time}
\title{Calculates the travel time from target cells across a friction
surface using \code{terra::costDist()}.}
\usage{
get_travel_time(target, friction)
}
\arguments{
\item{target}{A \code{\link[terra:SpatRaster]{terra::SpatRaster}} defining the target cells from which
travel time is calculated.}

\item{friction}{A \code{\link[terra:SpatRaster]{terra::SpatRaster}} representing the travel cost or
friction of each cell.#'}
}
\value{
A \code{\link[terra:SpatRaster]{terra::SpatRaster}} containing the accumulated least-cost travel
cost from the target cells.
}
\description{
Calculates the travel time from target cells across a friction
surface using \code{terra::costDist()}.
}
