% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Creates a regular grid of polygons covering an area.}
\usage{
make_grid(area, gridsize)
}
\arguments{
\item{area}{A \code{\link[terra:SpatVector]{terra::SpatVector}} defining the area to cover.}

\item{gridsize}{The grid cell size in map units.}
}
\value{
A \code{\link[terra:SpatVector]{terra::SpatVector}} containing the grid cells, with an \code{id}
attribute identifying each cell.
}
\description{
Creates a regular grid of polygons covering an area.
}
