% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_buildings.R
\name{retrieve_buildings}
\alias{retrieve_buildings}
\title{Retrieves OSM building footprint data for Edinburgh}
\usage{
retrieve_buildings()
}
\value{
A \code{\link[terra:SpatVector]{terra::SpatVector}}.
}
\description{
Retrieves OSM building footprint data for Edinburgh
}
