% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_dem.R
\name{retrieve_dem}
\alias{retrieve_dem}
\title{Retrieves AWS DEM data for Edinburgh}
\usage{
retrieve_dem()
}
\value{
A \code{\link[terra:SpatRaster]{terra::SpatRaster}}.
}
\description{
Retrieves AWS DEM data for Edinburgh
}
