% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_edinburgh.R
\name{retrieve_edinburgh}
\alias{retrieve_edinburgh}
\title{Retrieves the City of Edinburgh locality boundary}
\usage{
retrieve_edinburgh()
}
\value{
A \code{\link[terra:SpatVector]{terra::SpatVector}}.
}
\description{
Retrieves the City of Edinburgh locality boundary
}
