% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_roads.R
\name{retrieve_roads}
\alias{retrieve_roads}
\title{Retrieves OSM road network data for Edinburgh}
\usage{
retrieve_roads()
}
\value{
A \code{\link[terra:SpatVector]{terra::SpatVector}}.
}
\description{
Retrieves OSM road network data for Edinburgh
}
