\name{travelgrid-package}
\alias{travelgrid-package}
\alias{travelgrid}
\docType{package}
\title{Measures the time taken to travel somewhere through a landscape}
\description{
  A fast, scalable, package built to handle many-to-many least-cost travel
  time computations where the built environment and topography matter.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Chris Littleboy
Maintainer: Chris Littleboy <chris.littleboy@googlemail.com>
}
\keyword{ package }
